Reworked Java Implementation
2 New Packages:
---------------------------------------------------
src - This is the root package, includes:
* The GUI and all running VIzu-80 Components
* Initial (rough) implementation

src.cpu - The CPU Implementation, which includes:
* z80 CPU Implementation (version 0.0.a)
* Memory, Device, and CPU Interfaces
* Registers, Enums, and Exceptions
---------------------------------------------------

This is the initial commit and contains the updates mentioned above;
reworking the original implementation to work better with the visual
(GUI) aspect of this project.

Goals:
- Create a Visual Emulator, that's shows the internal states of the
system, while showing the effect of the OpCode on all internal states
and external devices.
- Used as a learning device to the internals of a CPU

Next Update:
- Version 0.1.a of the CPU Core; including a rough sketch of the OpCode
decoding process, without concrete implementations of each OpCode.